Complex Interactive System
Visualizing topic modeling, not an accurate display but an exploration.

This project serves as a visual sketchbook for my personal investigation into topic modeling and the relational nature of verbal exchange. I chose topics that elicit emotion in socially charged forms of media and presented them as a system that continuously released new data pointers "strings of characters" when a mouse is pressed. As the words appear on the screen, they continue to form connections, converging and diverging depending on their position.